--- src/main/java/nextstep/member/ui/ControllerExceptionHandler.java
@@ -1,41 +1,0 @@
-package nextstep.member.ui;
-
-import nextstep.member.ui.error.ErrorResponse;
-import nextstep.member.ui.error.exception.BusinessException;
-import org.springframework.dao.DataIntegrityViolationException;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ControllerAdvice;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-
-import java.util.List;
-
-@ControllerAdvice
-public class ControllerExceptionHandler {
-    @ExceptionHandler(DataIntegrityViolationException.class)
-    public ResponseEntity<Void> handleIllegalArgsException(DataIntegrityViolationException e) {
-        return ResponseEntity.badRequest().build();
-    }
-
-    /**
-     * 비즈니스 로직 실행 중 오류 발생
-     */
-    @ExceptionHandler(value = { BusinessException.class })
-    protected ResponseEntity<ErrorResponse> handleConflict(BusinessException e) {
-        List<String> errorMessages = List.of(e.getMessage());
-        HttpStatus httpStatus = e.getStatus();
-        ErrorResponse errorResponse = ErrorResponse.of(httpStatus, errorMessages);
-        return ResponseEntity.status(httpStatus).body(errorResponse);
-    }
-
-    /**
-     * 나머지 예외 발생
-     */
-    @ExceptionHandler(Exception.class)
-    protected ResponseEntity<ErrorResponse> handleException(Exception e) {
-        List<String> errorMessages = List.of(e.getMessage());
-        ErrorResponse errorResponse = ErrorResponse.of(HttpStatus.INTERNAL_SERVER_ERROR, errorMessages);
-        return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(errorResponse);
-    }
-
-}
